+++ bot/haiku/models/key.py
@@ -1,0 +1,9 @@
+from enum import Enum
+
+
+class HaikuKey(Enum):
+    """Partition keys for the haiku table."""
+
+    FIVE = "FIVE"
+    SEVEN = "SEVEN"
+    FORMED = "FORMED"
